// src/api/models/StoreRevenueReq.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StoreRevenueReq = {
    monthlyRevenue?: number;
    deliverySalesRatio?: number;
    maleCustomer2030Ratio?: number;
    maleCustomer40PlusRatio?: number;
    returningCustomerRatio?: number;
    year?: number;
    month?: number;
    storeId?: number;
};